move foundational types to base_types
It was becoming easy to introduce dependency cycles in generated code. While this
convention doesn't prevent cycles (eslint should already do that),
it hopefully makes it more clear which types can be used by generated
code.
