+++ libs/logging/src/base_types/log_event_types.ts
