+++ libs/logging/src/index.ts
@@ -1,9 +1,8 @@
 /* istanbul ignore file */
+export * from './base_types';
 export * from './export';
 export * from './helpers';
 export * from './log_event_ids';
-export * from './log_event_types';
-export * from './log_source';
 export * from './logger';
 export * from './test_utils';
 export * from './types';
