+++ libs/logging/src/test_utils.ts
@@ -1,5 +1,5 @@
 import { Logger } from './logger';
-import { LogSource } from './log_source';
+import { LogSource } from './base_types/log_source';
 
 export function fakeLogger(logSource: LogSource = LogSource.System): Logger {
   const logger = new Logger(logSource);
